Raise and relax graceful-fs dependency
It may not be relevant to the use that this package makes but `graceful-fs` had a potential memory leak fixed in 4.1.15. Additionally unless there is known to be something broken in a higher version it is nice to have relaxed version requirements (`^` or `~`). Relaxed versions reduces the number of duplicate versions yarn/npm needs to install.